Fix error in remove_ph_time_var
When bp_cal:remove_ph_time_var is enabled we do the following:
1. solve for a time-dependent normalised bandpass calibration
2. remove phase variations with time in the bpcal field by solving for a time-dependent gain calibration (applying the above bandpass on the fly)
3. solve for a time-independent normalised bandpass applying the above gain calibration on the fly.

Step 2 above was incorrectly done on the field config['gain_cal_flux'].get('field', 'bpcal'). Normally this field is be the flux calibrator. However, it should have been the bandpass calibrator. if the two are not the same the remove_ph_time_var procedure was incorrect.

This commit fixes this error.