Fixes bug that prevented common bary freq grid
Before imaging multiple .MS's in HI the pipeline regrids them all to a common barycentric frequency grid. Strictly, the common grid should be calculated as the maximum range of barycentric frequencies available to all MS's and for all time intervals; and with the poorest frequency resolution, i.e., the maximum Doppler-corrected channel width.

However, for the sake of speed the common barycentric frequency grid is calculated based on the Doppler corrections at midnight of the days when the data were taken for the various MS's.

This was causing issues when, during an observation, the proper barycentric frequency grid deviated significantly from the common grid. The result was that, behind the scene, CASA/mstransform would use a slightly different grid than requested for some MS's. The resulting frequency mismatch between MS's would drive WSClean crazy and result in non-sensical HI cubes.

This commit implements a (temporary?) fix to this problem. It consists in leaving the calculation of the common barycentric frequency grid as is, but then discard first and last channels to avoid the above issue. As long as the channel width is of 1 km/s or more (the maximum variation of the Doppler correction during 12 h) this will work.

In the long run we may consider calculating the common barycentric grid in a more accurate way by considering all relevant time intervals and not only the one at midnight of the day of observation of the various .MS's.